\name{make.bib}
\alias{make.bib}
\title{Produce BibTeX file including automatic keys.}
\usage{
  make.bib(..., file = NULL)
}
\arguments{
  \item{...}{packages}

  \item{file}{name of .bib file \code{NULL} suppresses
  writing of the file. \code{""} will output to stdout.}
}
\value{
  invisible bibentry list.
}
\description{
  Produce BibTeX file including automatic keys.
}
\author{
  Claudia Beleites
}

