% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{make_geteasytargets}
\title{Interface to `make` for package building}
\usage{
make(target = "all", path = "./", makefile = "Makefile")

make_geteasytargets(path = "./", makefile = "Makefile")
}
\arguments{
\item{target}{Makefile target, use \code{make_geteasytargets} to get a list of possible targets.}

\item{path}{path where makefile resides}

\item{makefile}{the Makefile to be used}
}
\description{
\code{make} calls make for the specified target.

\code{make_geteasytargets} extracts "easy" targets from the makefile, i.e.
non-complex (no "%") targets that are also not filenames.
}
\examples{
\dontrun{
make ("build")
}
\dontrun{
make_geteasytargets ()
}
}

